+++ vm/vm_item_pp2val.h
@@ -3,6 +3,6 @@
 
 #include "vm_stack.h"
 
-int stack_item_pp2value(stack_item* item);
+void stack_item_pp2value(stack_item* item);
 
 #endif